fixed a warning in liveapp unit test
git-svn-id: file:///var/www/production/sfweb/svn/trunk@419 ee427ae8-e902-0410-961c-c3ed070cd9f9

+++ test/liveapp/propelCrudGenerator/propelCrudGeneratorTest.php
@@ -30,7 +30,10 @@ class propelCrudGeneratorTest extends sfLiveProjectUnitTestCase
 
     // force autoload classes regeneration
     $autoload_config_file = sfConfig::get('sf_app_config_dir_name').'/autoload.yml';
-    unlink(sfConfigCache::getCacheName($autoload_config_file));
+    if (is_readable($autoload_config_file))
+    {
+      unlink(sfConfigCache::getCacheName($autoload_config_file));
+    }
     require(sfConfigCache::checkConfig($autoload_config_file));
 
     // create database
